organize flux code the same way the rest of lib is
